[keyframe] Handle corner case when identifying subsequences
This commit handles the specific case where the last frame of an input
sequence happens to be the one that pushes the motion accumulation over
the threshold: the last frame was being pushed a first time within the
'for' loop, and then a second time outside of it. This led the last
subsequence to have a size of 0, which led to the 'while' loop condition
filling the weights' vector to never be broken.